Modernize (parent pom and Jenkins) (#87)
Upgrading parent pom to be up-to-date.

Upgrading Jenkins as JCasC requires at least 2.60.3.

Fixing jdepend-maven-plugin version. There is no version 2.9.1. It looks
like this was a copy'n'paste error from maven-javadoc-plugin version.

Setting upper bound version for plugins via <dependencyManagement>
to be compliant with Maven enforcer rules.

Upgrading maven-plugin so that `mvn hpi:run` runs without errors.

Use SpotBugs instead of FindBugs in reporting (see change in parent pom).

Jenkins::getInstance is now annotated with @Nonnull. Therefore removing
the unnecessary null checks that have been detected by SpotBugs.

HtmlUnit removed/changed some methods, adapting tests to compile again.

Plugin "maven-plugin" is not required to be loaded explicitly in a test.
It is loaded automatically because it is defined as dependency in pom.xml.
Therefore just verifying that it is loaded is enough.