Reduce deprecated API usage (#116)
Now using `Jenking.get()` instead of `Jenkins.getInstance()`.

This new method is annotated with `@Nonnull` as it ensures that an
instance is returned. If none is available, an `IllegalStateException`
is thrown. This may simplify code as no `null` checks are required
anymore. It also means that 2 tests had to be converted to integration
tests to provide a Jenkins instance at runtime.