Remove unnecessary getJenkins() implementation
Beeing able to overwrite this was used for testing purposes only.
This complicated code a bit and was not used very often.

The only use case at the moment is to test authorization for some methods.
This can also be achieved by overwriting `getAccessControlledObject()`
which was already done in other tests.