Simplify code (#118)
* Use Jenkins instance from parent class

`JobConfigHistoryBaseAction` already contains a `Jenkins` instance.

* Remove unnecessary constructor implementation

If `nodes` is anyway set on first call of `onConfigurationChange` it is
not required to set it in the constructor. This simplifies the code.

* Remove unnecessary getJenkins() implementation

Beeing able to overwrite this was used for testing purposes only.
This complicated code a bit and was not used very often.

The only use case at the moment is to test authorization for some methods.
This can also be achieved by overwriting `getAccessControlledObject()`
which was already done in other tests.