WIP Re-enable accidentially disabled integration tests
The culprit is commit 958086294db1dc4f57e3d0f07bbc7fa4f68eaac9 when switching
from HudsonTestCase to JenkinsRule. The rule has to be used as a JUnit 4 rule
inside a test. It is different to JUnit 3 and cannot be used as a base class.
As no JUnit 3 base class is used anymore the JUnit 4 runner kicks in and only
detects methods annotated with `@Test` as test methods. JUnit 3 was using the
prefix "test" in method names to detect test methods.

The temporary folder has to be provided as absolute path.
Otherwise `Jenkins::checkRawBuildsDir` fails with a message like this:
"jenkins.model.InvalidBuildsDir: ${ITEM_ROOTDIR}/builds does not resolve to an absolute path"

`WebClient::goTo` now takes a relative instead of an absolute URL.

Triggering the "change" event on HTML form inputs now requires a bit
more plumbing.

Many tests fail because running the Javascript code fails reporting
a missing "each" function.