Implement workaround for JS exception in htmlunit
Apparently, htmlunit's WebClient has a problem executing either
dispatchEvent() or the fact, that this introduces a recursive call
of event handlers. This commit provides a workaround for that, by
disabling the invocation of dispatchEvent(). if a specia flag is
set in advance from within the test case. Consequently, the test code
has to invoke form.submit() itself.